Add simple access control to CMS modules and form for new admins
Except of that little corrections in directories structure

+++ fp-admin/php_scripts/menu/add_position.php
@@ -85,20 +85,24 @@
         // INSERT DATA INTO DATABASE
         if($OK == true)
         {
-            require_once "../../connect.php";
-            $sql = "INSERT INTO `menu_positions`(`position_id`, `position_name`, `position_reference_to`, `position_cat_reference`, `position_icon_path`) VALUES (NULL, ?, ?, ?, ?)";
-            $add = $conn->prepare($sql);
-            try
+            require_once  "../permissions/check.php";
+            if(check_site() == true)
             {
-                $add->execute([$position_name, $position_ref, $position_ref_cat, $new_image_name]);
-                $_SESSION['result'] = "Dodano pomyślnie!";
-            }
-            catch(Exception $e)
-            {
-                $_SESSION['result'] = "Wystąpił błąd!";
-                echo "<br>".$e;
-                unlink("../../img-db/menu_icons/".$new_image_path);
-            }
+                require_once "../../connect.php";
+                $sql = "INSERT INTO `menu_positions`(`position_id`, `position_name`, `position_reference_to`, `position_cat_reference`, `position_icon_path`) VALUES (NULL, ?, ?, ?, ?)";
+                $add = $conn->prepare($sql);
+                try
+                {
+                    $add->execute([$position_name, $position_ref, $position_ref_cat, $new_image_name]);
+                    $_SESSION['result'] = "Dodano pomyślnie!";
+                }
+                catch(Exception $e)
+                {
+                    $_SESSION['result'] = "Wystąpił błąd!";
+                    echo "<br>".$e;
+                    unlink("../../img-db/menu_icons/".$new_image_path);
+                }
+            }  
         }
         $conn = null;
         header("Location: ../../menu_editor.php");
